Attempt to support Abort/Kill on NET8
